remove most of the AddOn support
The dynamic loading nature of addons doesn't work well with the asset pipeline.
Totally removing it is safe, but it make is hard to use addons.
A more feasible solution for upstream would to keep the AddOn support,
and provide a way for it to work without the dynamic loading.